Map extra directories into the VM
I've got a Wordpress theme that I'd like to keep in a repository with
some non-Wordpress stuff. I want to use Chassis during development, but
I don't want to muck about with Chassis in my repository as well. I can
symlink `Chassis/content` or bits inside it all day long, but the VM
that Chassis creates has no idea what is on the other end of that
symlink.

So, I've added a tiny bit to the Vagrantfile that searches through the
coalesced config for an entry named `synced_folders`. It'll go through
the entire list, mapping the key (on the host filesystem) to the value
(on the virtual filesystem).