Make the WordPress site title configurable through yaml
While working on client projects it can be a nice touch to ensure a new
VM has the appropriate site title. There was a line of code in chassis
which seemed to try to set the installed name to 'Vagrant Site,' but it
was using the wrong wp::site parameter and it did not allow the value
to be configured. This patch adds a new `site` key to the YAML file,
permitting users to specity a `name` value to use as the site title for
their new VM.