Exclude parent directory from hyphenated class check
Fixes #689

A parent directory such as .../my-vm has no impact on Puppet's new
stricter rules around using "-" instead of "-" in class names, but
if Chassis was checked out into a hyphenated directory, our logic
previously detected and warned on the "-" in that parent directory
even if no hyphenated Puppet class was present.

If we do not exclude the parent directory from the string when testing
for or, more critically, creating new correct path strings, we end up
trying to clone the repositories into a non-existent directory path.