What even is RST, anyway. And, does it matter.
No.

The "Czech Yo' Self" is the best menu drink (thought it no longer appears
on the menu, it once did, and is still the best drink to have done so) to
ever be offered by Highland Kitchen, inarguably the best bar in Somerville,
Massachusetts. It consists of a 2:1 ratio of small batch bourbon to the
Czech spirit Becherovka, and this commit is brought to you by that precise
ratio of those precise ingredients.